+++ packages/doenetml-worker-javascript/src/StatePersistence.ts
@@ -1,0 +1,149 @@
+import { serializedComponentsReplacer, data_format_version } from "@doenet/utils";
+import { set as idb_set } from "idb-keyval";
+
+/**
+ * Owns the save-to-localStorage and save-to-database pipeline for a Core
+ * instance, including throttle timers and the debounced save scheduler.
+ *
+ * Holds a back-reference to Core to read `cumulativeStateVariableChanges`,
+ * `rendererState`, `flags`, `document`, the activity/doc/attempt IDs, and
+ * `coreInfoString` (set by Core during `generateDast`), and to invoke
+ * `reportScoreAndStateCallback`.
+ *
+ * This is purely the persistence I/O — the essential-value write engine
+ * that produces `cumulativeStateVariableChanges` is a separate concern
+ * (see `processNewStateVariableValues` in Core, slated for Phase 4).
+ */
+export class StatePersistence {
+    core: any;
+    saveStateToDBTimerId: ReturnType<typeof setTimeout> | null;
+    saveDocStateTimeoutID: ReturnType<typeof setTimeout> | null;
+    docStateToBeSavedToDatabase: any;
+    changesToBeSaved: boolean;
+
+    constructor({ core }: { core: any }) {
+        this.core = core;
+        this.saveStateToDBTimerId = null;
+        this.saveDocStateTimeoutID = null;
+        this.docStateToBeSavedToDatabase = null;
+        this.changesToBeSaved = false;
+    }
+
+    /**
+     * Schedule a debounced `saveState` after `delayMs` milliseconds, replacing
+     * any previously scheduled save.
+     */
+    scheduleSave(delayMs: number): void {
+        if (this.saveDocStateTimeoutID !== null) {
+            clearTimeout(this.saveDocStateTimeoutID);
+        }
+        this.saveDocStateTimeoutID = setTimeout(() => {
+            this.saveState();
+        }, delayMs);
+    }
+
+    async saveImmediately(): Promise<void> {
+        if (this.saveDocStateTimeoutID) {
+            // if in debounce to save doc to local storage
+            // then immediate save to local storage
+            // and override timeout to save to database
+            clearTimeout(this.saveDocStateTimeoutID);
+            await this.saveState(true);
+        } else {
+            // else override timeout to save any pending changes to database
+            await this.saveChangesToDatabase(true);
+        }
+    }
+
+    async saveState(
+        overrideThrottle = false,
+        onSubmission = false,
+    ): Promise<void> {
+        this.saveDocStateTimeoutID = null;
+
+        const core = this.core;
+
+        if (!core.flags.allowSaveState && !core.flags.allowLocalState) {
+            return;
+        }
+
+        let coreStateString = JSON.stringify(
+            core.cumulativeStateVariableChanges,
+            serializedComponentsReplacer,
+        );
+        let rendererStateString: string | null = null;
+
+        if (core.flags.saveRendererState) {
+            rendererStateString = JSON.stringify(
+                core.rendererState,
+                serializedComponentsReplacer,
+            );
+        }
+
+        if (core.flags.allowLocalState) {
+            await idb_set(
+                `${core.activityId}|${core.docId}|${core.attemptNumber}|${core.cid}`,
+                {
+                    data_format_version,
+                    coreState: coreStateString,
+                    rendererState: rendererStateString,
+                    coreInfo: core.coreInfoString,
+                },
+            );
+        }
+
+        if (!core.flags.allowSaveState) {
+            return;
+        }
+
+        this.docStateToBeSavedToDatabase = {
+            cid: core.cid,
+            coreInfo: core.coreInfoString,
+            coreState: coreStateString,
+            rendererState: rendererStateString,
+            initializeCounters: core.initializeCounters,
+            docId: core.docId,
+            attemptNumber: core.attemptNumber,
+            activityId: core.activityId,
+            onSubmission,
+        };
+
+        // mark presence of changes
+        // so that next call to saveChangesToDatabase will save changes
+        this.changesToBeSaved = true;
+
+        // if not currently in throttle, save changes to database
+        await this.saveChangesToDatabase(overrideThrottle);
+    }
+
+    async saveChangesToDatabase(overrideThrottle = false): Promise<void> {
+        // throttle save to database at 60 seconds
+
+        if (!this.changesToBeSaved) {
+            return;
+        }
+
+        if (this.saveStateToDBTimerId !== null) {
+            if (overrideThrottle) {
+                clearTimeout(this.saveStateToDBTimerId);
+            } else {
+                return;
+            }
+        }
+
+        this.changesToBeSaved = false;
+
+        // check for changes again after 60 seconds
+        this.saveStateToDBTimerId = setTimeout(() => {
+            this.saveStateToDBTimerId = null;
+            this.saveChangesToDatabase();
+        }, 60000);
+
+        this.core.reportScoreAndStateCallback({
+            state: { ...this.docStateToBeSavedToDatabase },
+            score: await this.core.document.stateValues.creditAchieved,
+        });
+
+        return;
+    }
+}
